aggregator/requirements.txt
pyyaml>=6.0
openai>=1.0
# gitpython is no longer required; cloner uses git CLI directly
